+++ src/js/versionInfo.js
@@ -3,5 +3,5 @@
  *           without modifying formatting for automation scripts
  * @type {{VERSION: string}}
  */
-let VersionInfo = { VERSION : '0.0.1485' };
+let VersionInfo = { VERSION : '0.0.1495' };
 export default VersionInfo
